LayoutTests/editing/selection/skip-over-uneditable-in-contenteditable-expected.txt

Before

Middle

After

Ensure that extending a selection inside a contentEditable skips past an uneditable region.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS sel.focusNode is after
PASS sel.focusNode.parentElement is before
PASS sel.focusNode is after
PASS sel.focusNode.parentElement is before
PASS sel.focusNode.parentElement is after
PASS sel.focusNode.parentElement is before
PASS successfullyParsed is true

TEST COMPLETE
